requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
